JAMES-1877 Extract DeliveryRunnable from RemoteDelivery
